move mdc id to preHandle
