--- src/main/java/bio/terra/workspace/mdc/MdcResponseFilter.java
@@ -1,24 +1,0 @@
-package bio.terra.workspace.mdc;
-
-import java.io.IOException;
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.slf4j.MDC;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-@Component
-public class MdcResponseFilter extends OncePerRequestFilter {
-
-  @Override
-  protected void doFilterInternal(
-      HttpServletRequest httpServletRequest,
-      HttpServletResponse httpServletResponse,
-      FilterChain filterChain)
-      throws ServletException, IOException {
-    httpServletResponse.addHeader("X-Request-ID", MDC.get("requestId"));
-    filterChain.doFilter(httpServletRequest, httpServletResponse);
-  }
-}
